# docs/swagger.yaml
basePath: /
host: localhost:3001
info:
  contact:
    email: ppreeper@gmail.com
    name: Peter Preeper
  description: This is the swagger spec for AD Dict
  license:
    name: AGPL-3
    url: https://www.gnu.org/licenses/agpl-3.0.html
  termsOfService: http://preeper.org/addict/
  title: AD Dict API
  version: "1.0"
paths:
  /api/all:
    get:
      consumes:
      - text/plain
      description: Pulls all Active Directory objects
      parameters:
      - collectionFormat: csv
        description: Comma-delimited field names to return
        in: query
        items:
          type: string
        name: _fields
        type: array
      - collectionFormat: csv
        description: Searches all fields for given string
        in: query
        items:
          type: string
        name: _q
        type: array
      - description: Result Index to start from
        in: query
        name: _start
        type: integer
      - description: Result Index to end to
        in: query
        name: _end
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get all objects
      tags:
      - other
  /api/find/{filter}:
    get:
      consumes:
      - text/plain
      description: Does a raw Active Directory search
      parameters:
      - description: Search filter
        in: path
        name: filter
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Search Active Directory
      tags:
      - other
  /api/group:
    get:
      consumes:
      - text/plain
      description: Pulls all groups in Active Directory, with filters.
      parameters:
      - collectionFormat: csv
        description: Comma-delimited field names to return
        in: query
        items:
          type: string
        name: _fields
        type: array
      - collectionFormat: csv
        description: Searches all fields for given string
        in: query
        items:
          type: string
        name: _q
        type: array
      - description: Result Index to start from
        in: query
        name: _start
        type: integer
      - description: Result Index to end to
        in: query
        name: _end
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get all groups
      tags:
      - group
    post:
      description: Adds a new group to Active Directory
      parameters:
      - description: Name of the group as displayed
        in: formData
        name: cn
        required: true
        type: string
      - description: Descripton of the group
        in: formData
        name: description
        type: string
      - description: Relative AD Position separated by / (ex Users or Users/Office)
        in: formData
        name: location
        required: true
        type: string
      responses:
        "201":
          description: Created
      summary: Add a group
      tags:
      - group
  /api/group/{group}:
    delete:
      consumes:
      - text/plain
      description: Removes a group from Active Directory
      parameters:
      - description: The group name
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Remove a group
      tags:
      - group
    get:
      consumes:
      - text/plain
      description: Pulls a single group
      parameters:
      - description: The group name
        in: path
        name: group
        required: true
        type: string
      - collectionFormat: csv
        description: Comma-delimited field names to return
        in: query
        items:
          type: string
        name: _fields
        type: array
      - collectionFormat: csv
        description: Searches all fields for given string
        in: query
        items:
          type: string
        name: _q
        type: array
      - description: Result Index to start from
        in: query
        name: _start
        type: integer
      - description: Result Index to end to
        in: query
        name: _end
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get a single group
      tags:
      - group
  /api/group/{group}/exists:
    get:
      consumes:
      - text/plain
      description: Returns whether a group exists or not
      parameters:
      - description: The group name
        in: path
        name: group
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Group exists
      tags:
      - group
  /api/group/{group}/user/{user}:
    delete:
      consumes:
      - text/plain
      description: Removes a user from a group
      parameters:
      - description: The group name
        in: path
        name: group
        required: true
        type: string
      - description: The user logon name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Remove user from group
      tags:
      - group
    post:
      consumes:
      - text/plain
      description: Adds a user to a group
      parameters:
      - description: The group name
        in: path
        name: group
        required: true
        type: string
      - description: The user logon name
        in: path
        name: user
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Add user to group
      tags:
      - group
  /api/other:
    get:
      description: Pulls all non-user/group Active Directory objects
      parameters:
      - collectionFormat: csv
        description: Comma-delimited field names to return
        in: query
        items:
          type: string
        name: _fields
        type: array
      - collectionFormat: csv
        description: Searches all fields for given string
        in: query
        items:
          type: string
        name: _q
        type: array
      - description: Result Index to start from
        in: query
        name: _start
        type: integer
      - description: Result Index to end to
        in: query
        name: _end
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get all other objects
      tags:
      - other
  /api/ou:
    get:
      consumes:
      - text/plain
      description: Pulls all Organization Units in Active Directory, with filters
      parameters:
      - collectionFormat: csv
        description: Comma-delimited field names to return
        in: query
        items:
          type: string
        name: _fields
        type: array
      - collectionFormat: csv
        description: Searches all fields for given string
        in: query
        items:
          type: string
        name: _q
        type: array
      - description: Result Index to start from
        in: query
        name: _start
        type: integer
      - description: Result Index to end to
        in: query
        name: _end
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get all OUs
      tags:
      - ou
    post:
      consumes:
      - text/plain
      description: Adds a new OU to Active Directory
      parameters:
      - description: Name of the OU as displayed
        in: body
        name: name
        required: true
        schema:
          type: string
      - description: Descripton of the OU
        in: body
        name: description
        schema:
          type: string
      - description: Relative AD Position separated by /
        in: body
        name: location
        schema:
          type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
      summary: Add an OU
      tags:
      - ou
  /api/ou/{ou}:
    delete:
      consumes:
      - text/plain
      description: Removes the OU from Active Directory
      parameters:
      - description: The OU name
        in: path
        name: ou
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Remove an OU
      tags:
      - ou
    get:
      consumes:
      - text/plain
      description: Pulls a single OU
      parameters:
      - description: The OU name
        in: path
        name: ou
        required: true
        type: string
      - collectionFormat: csv
        description: Comma-delimited field names to return
        in: query
        items:
          type: string
        name: _fields
        type: array
      - collectionFormat: csv
        description: Searches all fields for given string
        in: query
        items:
          type: string
        name: _q
        type: array
      - description: Result Index to start from
        in: query
        name: _start
        type: integer
      - description: Result Index to end to
        in: query
        name: _end
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: Get a single OU
      tags:
      - ou
  /api/ou/{ou}/exists:
    get:
      consumes:
      - text/plain
      description: Returns whether a OU exists or not
      parameters:
      - description: The OU name
        in: path
        name: ou
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: OU exists
      tags:
      - ou
...
swagger: "2.0"